resolve comments - build ece file detail
